doc: Remove struct/union/enum, add type keyword
I did this because it's hard to distinguish between the tuple type and
tuple value. If we force tuple type to be combined with `struct` like
`struct (i32, i32)` and value to be `.(3, 4)` with dot ahead, then it
would be too verbose.

So I decide to use `type (i32, i32)` to represent the tuple type,
otherwise it's tuple value.